Update README.md
This description is the README.md file. Such a file can be created automatically when creating a new repository (only the project title and optional description is inserted), but you can also create it manually at any time. Such a file has the .md extension, which means that its content is written using the MarkDown syntax.